fix: admin auth accessable by users accidentally
