+++ fe/app/(authenticated)/admin/page.tsx
@@ -1,249 +1,9 @@
-"use client";
+import { requireStaffRole } from "@/lib/auth-guard";
 
-import {
-  AlertCircle,
-  Edit,
-  RotateCcw,
-  Search,
-  Shield,
-  UserPlus,
-  Users,
-  X,
-} from "lucide-react";
-import { useCallback, useMemo, useState } from "react";
+import AdminPageClient from "./page.client";
 
-import CreateUserModal from "@/components/new/forms/admin/CreateUserModal";
-import EditUserRolesModal from "@/components/new/forms/admin/EditUserRolesModal";
-import { useUsers } from "@/hooks/users/useUsers";
-import { mappedRole } from "@/lib/utils";
-import { User } from "@/types/user";
+export default async function AdminPage() {
+  await requireStaffRole();
 
-export default function AdminPage() {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [selectedUser, setSelectedUser] = useState<User | null>(null);
-  const [showEditModal, setShowEditModal] = useState(false);
-  const [showCreateModal, setShowCreateModal] = useState(false);
-
-  const { data: users = [], isLoading, refetch, error } = useUsers();
-
-  const filteredUsers = useMemo(() => {
-    if (!searchQuery.trim()) return users;
-
-    const query = searchQuery.toLowerCase().trim();
-    return users.filter(
-      (user) =>
-        user.id.toString().includes(query) ||
-        user.username.toLowerCase().includes(query) ||
-        user.firstName.toLowerCase().includes(query) ||
-        user.lastName.toLowerCase().includes(query) ||
-        user.roles.some((role) =>
-          mappedRole(role).toLowerCase().includes(query)
-        )
-    );
-  }, [users, searchQuery]);
-
-  const handleSearchChange = useCallback((value: string) => {
-    setSearchQuery(value);
-  }, []);
-
-  const handleClearSearch = useCallback(() => {
-    setSearchQuery("");
-  }, []);
-
-  const handleEditUser = useCallback((user: User) => {
-    setSelectedUser(user);
-    setShowEditModal(true);
-  }, []);
-
-  const handleCreateUser = useCallback(() => {
-    setShowCreateModal(true);
-  }, []);
-
-  const handleCloseModal = useCallback(() => {
-    setSelectedUser(null);
-    setShowEditModal(false);
-    setShowCreateModal(false);
-  }, []);
-
-  const handleSuccess = useCallback(() => {
-    handleCloseModal();
-    refetch();
-  }, [handleCloseModal, refetch]);
-
-  const getRoleColor = (role: string) => {
-    const colors: Record<string, string> = {
-      Admin: "bg-red-100 text-red-800",
-      Doctor: "bg-blue-100 text-blue-800",
-      Patient: "bg-purple-100 text-purple-800",
-      Staff: "bg-green-100 text-green-800",
-    };
-    return colors[role] || "bg-gray-100 text-gray-800";
-  };
-
-  if (isLoading) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <div className="text-center">
-          <div className="mb-2 animate-spin text-blue-500">
-            <Users className="size-10" />
-          </div>
-          <p className="text-gray-600">Loading users...</p>
-        </div>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <div className="text-center">
-          <AlertCircle className="mx-auto mb-4 size-12 text-red-500" />
-          <p className="mb-4 text-gray-700">Failed to load users.</p>
-          <button
-            onClick={() => refetch()}
-            className="flex items-center gap-2 rounded bg-blue-600 px-4 py-2 text-white"
-          >
-            <RotateCcw className="size-4" />
-            Retry
-          </button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-white p-6">
-      <div className="mx-auto max-w-7xl">
-        <div className="mb-6">
-          <div className="mb-4 flex items-center justify-between">
-            <div className="flex items-center gap-4">
-              <div className="rounded-xl bg-blue-600 p-3">
-                <Shield className="size-8 text-white" />
-              </div>
-              <div>
-                <h1 className="text-3xl font-bold text-gray-800">
-                  User Management
-                </h1>
-                <p className="text-gray-600">
-                  Manage system users and their roles
-                </p>
-              </div>
-            </div>
-            <button
-              onClick={handleCreateUser}
-              className="flex items-center gap-2 rounded-lg bg-green-600 px-4 py-2 text-sm font-medium text-white hover:bg-green-700"
-            >
-              <UserPlus className="size-4" />
-              Create New User
-            </button>
-          </div>
-
-          <div className="relative mb-4">
-            <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-4">
-              <Search className="size-5 text-gray-400" />
-            </div>
-            <input
-              type="text"
-              placeholder="Search by ID, email, name, or role..."
-              className="w-full rounded-xl border border-gray-200 bg-gray-50 px-12 py-3 text-gray-700 focus:border-blue-400 focus:bg-white focus:outline-none focus:ring-2 focus:ring-blue-200"
-              value={searchQuery}
-              onChange={(e) => handleSearchChange(e.target.value)}
-            />
-            {searchQuery && (
-              <button
-                onClick={handleClearSearch}
-                className="absolute inset-y-0 right-0 flex items-center pr-4"
-              >
-                <X className="size-5 text-gray-400 hover:text-gray-600" />
-              </button>
-            )}
-          </div>
-        </div>
-
-        <div className="overflow-hidden rounded-xl bg-white shadow-lg">
-          <div className="overflow-x-auto">
-            <table className="w-full">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th className="px-6 py-3 text-left text-xs font-medium uppercase tracking-wider text-gray-500">
-                    User
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium uppercase tracking-wider text-gray-500">
-                    Email
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium uppercase tracking-wider text-gray-500">
-                    Roles
-                  </th>
-                  <th className="px-6 py-3 text-right text-xs font-medium uppercase tracking-wider text-gray-500">
-                    Actions
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="divide-y divide-gray-200 bg-white">
-                {filteredUsers.map((user: User) => (
-                  <tr key={user.id} className="hover:bg-gray-50">
-                    <td className="whitespace-nowrap px-6 py-4">
-                      <div>
-                        <div className="text-sm font-medium text-gray-900">
-                          {user.firstName} {user.lastName}
-                        </div>
-                        <div className="text-sm text-gray-500">
-                          ID: {user.id}
-                        </div>
-                      </div>
-                    </td>
-                    <td className="whitespace-nowrap px-6 py-4 text-sm text-gray-900">
-                      {user.username}
-                    </td>
-                    <td className="whitespace-nowrap px-6 py-4">
-                      <div className="flex flex-wrap gap-1">
-                        {user.roles.map((role) => (
-                          <span
-                            key={role}
-                            className={`inline-flex min-w-[70px] justify-center rounded-full px-2 py-1 text-xs font-medium ${getRoleColor(
-                              mappedRole(role)
-                            )}`}
-                          >
-                            {mappedRole(role)}
-                          </span>
-                        ))}
-                      </div>
-                    </td>
-                    <td className="whitespace-nowrap px-6 py-4 text-right text-sm font-medium">
-                      <button
-                        onClick={() => handleEditUser(user)}
-                        className="inline-flex items-center gap-1 rounded-md bg-blue-600 px-3 py-1 text-xs text-white hover:bg-blue-700"
-                      >
-                        <Edit className="size-3" />
-                        Edit Roles
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-
-          {filteredUsers.length === 0 && (
-            <div className="py-16 text-center text-gray-600">
-              <Users className="mx-auto mb-4 size-12 text-gray-400" />
-              No users found.
-            </div>
-          )}
-        </div>
-      </div>
-
-      {showEditModal && selectedUser && (
-        <EditUserRolesModal
-          user={selectedUser}
-          onClose={handleCloseModal}
-          onSuccess={handleSuccess}
-        />
-      )}
-
-      {showCreateModal && (
-        <CreateUserModal onClose={handleCloseModal} onSuccess={handleSuccess} />
-      )}
-    </div>
-  );
+  return <AdminPageClient />;
 }
